CiAO IP-Stack: fixed most compiler warnings/errors on mspgcc4
git-svn-id: https://www4.informatik.uni-erlangen.de:8088/i4ciao/ciaoos/trunk@1392 857cdeed-beef-0310-bd7e-d983e88953ba

+++ tcp/TCP_Socket.cpp
@@ -215,7 +215,7 @@ void TCP_Socket::setMSS(unsigned max_segment_size) {
     UInt64 maxRecvWnd = ((UInt64)maxReceiveWindow_MSS) * max_segment_size;
     if(maxRecvWnd > 0xFFFFFFFFUL){
       //limit to 32 bit (4 GByte)
-      maxReceiveWindow_Bytes = 0xFFFFFFFFUL;
+      maxReceiveWindow_Bytes = (unsigned) 0xFFFFFFFFUL;
     }
     else{
       maxReceiveWindow_Bytes = (unsigned) maxRecvWnd;
